+++ CHANGELOG.md
@@ -31,6 +31,7 @@ Misc:
 - Allow `location(start_line,start_column)` format [#1257](https://github.com/sider/runners/pull/1257)
 - **ESLint** Add `target` option instead of `dir` option [#1264](https://github.com/sider/runners/pull/1264)
 - **HAML-Lint** Add `target` option [#1265](https://github.com/sider/runners/pull/1265)
+- **ESLint** Update the default configuration [#1270](https://github.com/sider/runners/pull/1270)
 
 ## 0.29.3